Suppress unittest
unittests requires access to didweb url which has changed,

Suppress the test so that other work can move forward.

Ticket raised to fix this error.

Additionally pip-audit found a vulnerability and the requests dependency
is upgraded to 2.32.0.
python-ecdsa GHSA-wj6h-64fc-37mp vulnerability is ignored as maintainres
will not fix it.